backend/requirements.txt
acres==0.3.0
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
blis==0.7.11
boto3==1.28.53
botocore==1.31.85
cachetools==5.5.2
camelot-py==0.11.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.18.0
Flask==3.1.0
Flask-Cors==3.0.10
fonttools==4.56.0
fsspec==2025.3.0
ghostscript==10.05.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==1.3.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.17.3
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jpype1==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
loguru==0.7.2
looseversion==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.31.0
neo4j==5.28.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.8.1
numpy==1.26.4
openai==1.60.1
opencv-python==4.11.0.86
openpyxl==3.1.2
packaging==23.2
pandas==2.2.3
pathlib==1.0.1
pdfminer==20191125
pdfminer.six==20221105
pillow>=8.3.2,<11.0.0
pluggy==1.5.0
postgrest==0.11.0
preshed==3.0.9
protobuf==4.25.6
prov==2.0.1
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.8.0
PyMuPDF==1.25.4
pyparsing==3.2.1
pypdf==4.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-jose==3.3.0
pytils==0.4.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
realtime==1.0.6
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.6.2
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.3.0
scipy==1.15.2
sentence-transformers==2.7.0
simplejson==3.20.1
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
storage3==0.6.1
streamlit==1.32.0
StrEnum==0.4.15
supabase==1.1.0
supafunc==0.2.3
sympy==1.13.1
tabula-py==2.8.2
tabulate==0.9.0
tenacity==8.5.0
thinc==8.2.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tokenizers==0.14.1
toml==0.10.2
tools==0.1.9
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.34.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.0.7
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.3.4
websockets==12.0
Werkzeug==3.1.3
pyxlsb==1.0.10
